CompactBases.jl

CompactBases.jl is a package for various bases for function approximation with compact support in the framework of ContinuumArrays

CompactBases.AbstractKnotSet

CompactBases.ArbitraryKnotSet

CompactBases.ArbitraryKnotSet

CompactBases.BSpline

CompactBases.BSpline

CompactBases.BSpline

CompactBases.ExpKnotSet

CompactBases.ExpKnotSet

CompactBases.LinearKnotSet

CompactBases.LinearKnotSet

CompactBases.RadialDifferences

CompactBases.UnitVector

Base.first

Base.first

Base.getindex

Base.getindex

Base.last

Base.last

Base.length

Base.length

CompactBases.assert_multiplicities

CompactBases.assert_multiplicities

CompactBases.change_interval!

CompactBases.deBoor

CompactBases.find_interval

CompactBases.find_interval

CompactBases.lagrangeder!

CompactBases.leftmultiplicity

CompactBases.leftmultiplicity

CompactBases.lgwt

CompactBases.lgwt

CompactBases.nonempty_intervals

CompactBases.nonempty_intervals

CompactBases.num_quadrature_points

CompactBases.num_quadrature_points

CompactBases.numfunctions

CompactBases.numfunctions

CompactBases.numintervals

CompactBases.numintervals

CompactBases.order

CompactBases.order

CompactBases.rightmultiplicity

CompactBases.rightmultiplicity

CompactBases.within_interval

CompactBases.within_interval

CompactBases.within_support

CompactBases.within_support

CompactBases.@materialize

CompactBases.ArbitraryKnotSet — Type

ArbitraryKnotSet{k,ml,mr}(t)

An arbitrary knot set of order k and left and right multiplicities of ml and mr, respectively. The knot set is specified by the non-decreasing vector t; the same knot may appear multiple times, which influences the continuity of the B-splines at that location.

CompactBases.ArbitraryKnotSet — Type

ArbitraryKnotset(k, t[, ml=k, mr=k])

Construct an order-k arbitrary knot set, with the locations of the knots given by non-decreasing vector t.

CompactBases.BSpline — Type

BSpline(t, x, w, B, S)

Basis structure for the B-splines generated by the knot set t. x and w are the associated quadrature roots and weights, respectively, the columns of B correspond to the B-splines resolved on the quadrature roots, and S is the (banded) B-spline overlap matrix.

CompactBases.BSpline — Method

BSpline(t[, N])

Create the B-spline basis corresponding to the knot set t. N is the amount of Gauß–Legendre quadrature points per interval.

CompactBases.BSpline — Method

BSpline(t[; k′=3])

Create the B-spline basis corresponding to the knot set t. k′ is the highest polynomial order of operators for which it should be possible to compute the matrix elements exactly (via Gauß–Legendre quadrature). The default k′=3 corresponds to operators O(x²).

CompactBases.ExpKnotSet — Type

ExpKnotSet{k,ml,mr}(exponents, base, t, include0)

A knot set of order k and left and right multiplicities of ml and mr, respectively, whose knots are exponentially distributed according to t = base .^ exponents, optionally including 0 as the left endpoint.

CompactBases.ExpKnotSet — Method

ExpKnotSet(k, a, b, N[, ml=k, mr=k, base=10, include0=true])

Construct an order-k knot spanning from base^a to base^b in N intervals, optionally including 0 as the left endpoint.

Examples

julia> ExpKnotSet(2, -4, 2, 7)
10-element ExpKnotSet{2,2,2,Float64,StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}},Array{Float64,1}}:
   0.0
   0.0
   0.0001
   0.001
   0.01
   0.1
   1.0
  10.0
 100.0
 100.0

CompactBases.LinearKnotSet — Type

LinearKnotSet{k,ml,mr}(t)

A knot set of order k and left and right multiplicities of ml and mr, respectively, whose knots are uniformly distributed according to the range t. The interior basis functions are thus Cᵏ⁻²-continuous.

CompactBases.LinearKnotSet — Type

LinearKnotSet(k, a, b, N[, ml=k, mr=k])

Construct an order-k linear knot set spanning from a to b, with N intervals.

Examples

julia> LinearKnotSet(2, 0, 1, 3)
6-element LinearKnotSet{2,2,2,Float64,StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}:
 0.0
 0.0
 0.3333333333333333
 0.6666666666666666
 1.0
 1.0

CompactBases.RadialDifferences — Method

RadialDifferences(rₘₐₓ::T, n::I, args...)

Convenience constructor for RadialDifferences covering the open interval (0,rₘₐₓ) with n grid points.

CompactBases.nonempty_intervals — Method

nonempty_intervals(t)

Return the indices of all intervals of the knot set t that are non-empty.

Examples

julia> nonempty_intervals(ArbitraryKnotSet(3, [0.0, 1, 1, 3, 4, 6], 1, 3))
4-element Array{Int64,1}:
 1
 3
 4
 5

CompactBases.numfunctions — Method

numfunctions(t)

Returns the number of basis functions generated by knot set t.

Examples

julia> numfunctions(LinearKnotSet(3, 0, 1, 2))
4

julia> numfunctions(LinearKnotSet(5, 0, 1, 2))
6

CompactBases.numintervals — Method

numinterval(t)

Returns the number of intervals generated by the knot set t.

Examples

julia> numintervals(LinearKnotSet(3, 0, 1, 2))
2

CompactBases.order — Method

order(t)

Returns the order k of the knot set t.

CompactBases.AbstractKnotSet — Type

AbstractKnotSet{k,ml,mr,T}

Abstract base for B-spline knot sets. T is the eltype of the knot set, k is the order of the piecewise polynomials (order = degree + 1) and ml and mr are the knot multiplicities of the left and right endpoints, respectively.

CompactBases.UnitVector — Type

UnitVector{T}(N, k)

Helper vector type of length N where the kth element is one(T) and all the others zero(T).

Base.first — Method

first(t)

Return the first knot of t.

Base.getindex — Method

getindex(t, i)

Return the ith knot of the knot set t, accounting for the multiplicities of the endpoints.

Examples

julia> LinearKnotSet(3, 0, 1, 3)[2]
0.0

julia> LinearKnotSet(3, 0, 1, 3, 1, 1)[2]
0.3333333333333333

Base.last — Method

last(t)

Return the last knot of t.

Base.length — Method

length(t)

Return the number of knots of t.

Examples

julia> length(LinearKnotSet(3, 0, 1, 3))
8

julia> length(LinearKnotSet(3, 0, 1, 3, 1, 1))
4

CompactBases.assert_multiplicities — Method

assert_multiplicities(k,ml,mr,t)

Assert that the multiplicities at the endpoints, ml and mr, respectively, are consistent with the order k. Also check that the amount of knots in the knot set t are enough to support the requested order k.

CompactBases.change_interval! — Method

change_interval!(x, w, xs, ws[, a=0, b=1])

Transform the Gaußian quadrature roots x and weights w on the elementary interval [-1,1] to the interval [a,b] and store the result in xs and ws, respectively.

CompactBases.deBoor — Method

deBoor(t, c, x[, i[, m=0]])

Evaluate the spline given by the knot set t and the set of control points c at x using de Boor's algorithm. i is the index of the knot interval containing x. If m≠0, calculate the mth derivative at x instead.

CompactBases.find_interval — Method

find_interval(t, x[, i=ml])

Find the interval in the knot set t that includes x, starting from interval i (which by default is the first non-zero interval of the knot set). The search complexity is linear, but by storing the result and using it as starting point for the next call to find_interval, the knot set need only be traversed once.

CompactBases.lagrangeder! — Method

lagrangeder!(xⁱ, m, L′)

Calculate the derivative of the Lagrange interpolating polynomial Lⁱₘ(x), given the roots xⁱ, at the roots, and storing the result in L′.

∂ₓ Lⁱₘ(xⁱₘ,) = (xⁱₘ-xⁱₘ,)⁻¹ ∏(k≠m,m′) (xⁱₘ,-xⁱₖ)/(xⁱₘ-xⁱₖ), m≠m′, [δ(m,n) - δ(m,1)]/2wⁱₘ, m=m′

Eq. (20) Rescigno2000

CompactBases.leftmultiplicity — Method

leftmultiplicity(t)

Return the multiplicity of the left endpoint.

CompactBases.lgwt — Method

lgwt(t, N) -> (x,w)

Generate the N Gauß–Legendre quadrature roots x and associated weights w, with respect to the B-spline basis generated by the knot set t.

Examples

julia> CompactBases.lgwt(LinearKnotSet(2, 0, 1, 3), 2)
([0.0704416, 0.262892, 0.403775, 0.596225, 0.737108, 0.929558], [0.166667, 0.166667, 0.166667, 0.166667, 0.166667, 0.166667])

julia> CompactBases.lgwt(ExpKnotSet(2, -4, 2, 7), 2)
([2.11325e-5, 7.88675e-5, 0.000290192, 0.000809808, 0.00290192, 0.00809808, 0.0290192, 0.0809808, 0.290192, 0.809808, 2.90192, 8.09808, 29.0192, 80.9808], [5.0e-5, 5.0e-5, 0.00045, 0.00045, 0.0045, 0.0045, 0.045, 0.045, 0.45, 0.45, 4.5, 4.5, 45.0, 45.0])

CompactBases.num_quadrature_points — Method

num_quadrature_points(k, k′)

The number of quadrature points needed to exactly compute the matrix elements of an operator of polynomial order k′ with respect to a basis of order k.

CompactBases.rightmultiplicity — Method

rightmultiplicity(t)

Return the multiplicity of the right endpoint.

CompactBases.within_interval — Method

within_interval(x, interval)

Return the indices of the elements of x that lie within the given interval.

CompactBases.within_support — Method

within_support(x, t, j)

Return the indices of the elements of x that lie withing the compact support of the jth basis function (enumerated 1..n), given the knot set t. For each index of x that is covered, the index k of the interval within which x[i] falls is also returned.

CompactBases.@materialize — Macro

@materialize function op(args...)

This macro simplifies the setup of a few functions necessary for the materialization of Applied objects:

ApplyStyle, used to ensure dispatch of the applied object to the routines below

copyto!(dest::DestType, applied_obj::Applied{...,op}) performs the actual materialization of applied_obj into the destination object which has been generated by

similar which usually returns a suitable matrix

materialize which makes use of the above functions

Example

@materialize function *(Ac::MyAdjointBasis,
                        O::MyOperator,
                        B::MyBasis)
    MyApplyStyle # An instance of this type will be returned by ApplyStyle
    T -> begin # generates similar
        A = parent(Ac)
        parent(A) == parent(B) ||
            throw(ArgumentError("Incompatible bases"))

        # There may be different matrices best representing different
        # situations:
        if ...
            Diagonal(Vector{T}(undef, size(B,1)))
        else
            Tridiagonal(Vector{T}(undef, size(B,1)-1),
                        Vector{T}(undef, size(B,1)),
                        Vector{T}(undef, size(B,1)-1))
        end
    end
    dest::Diagonal{T} -> begin # generate copyto!(dest::Diagonal{T}, ...) where T
        dest.diag .= 1
    end
    dest::Tridiagonal{T} -> begin # generate copyto!(dest::Tridiagonal{T}, ...) where T
        dest.dl .= -2
        dest.ev .= 1
        dest.du .= 3
    end
end